# Optimize the exFAT free disk space calculator