@GideonZ 2025-01-07
Maybe it would be good to notify the maker of ChanFAT about this. This looks like excellent work!
I haven't tested it, though. How did you verify its behavior?